requirements.txt
absl-py==1.4.0
albumentations==1.3.1
arrow==1.2.3
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
Deprecated==1.2.14
efficientnet-pytorch==0.6.3
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.6.0
gast==0.4.0
glob2==0.7
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
h5py==3.8.0
huggingface-hub==0.15.1
humanize==4.6.0
idna==3.4
imageio==2.31.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
monai==1.2.0
mpmath==1.3.0
munch==3.0.0
networkx==3.1
nibabel==5.1.0
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pretrainedmodels==0.7.4
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-time==0.3.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
random2==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
segmentation-models-pytorch==0.2.0
shellingham==1.5.0.post1
SimpleITK==2.2.1
six==1.16.0
sympy==1.12
temp==2020.7.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.4.12
torch==2.0.1
torchio==0.18.91
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.15.0
